deepen: absorb extractor utils and delete debug fork into Filing extractor
Replaces 3 shallow modules with one deep Filing intake extractor module
behind a 2-method interface (extractFilingContent, cleanHtmlContent).

Locality: every line of Filing-content extraction now lives in
lib/parsers/filing-extractor.ts. Constants (SEC_BASE_URL, MAX_CONTENT_SIZE,
PATTERNS, PRIORITY_EXTENSIONS), the worker-pool, HTML/XML cleaners,
directory-listing detection, document-link extraction, and URL resolution
all collapse from cross-file imports into private helpers behind the same
interface.

Leverage: callers reach for one module. The only external user of the
utils module (services/filing/summaryGenerationService.ts) updates one
import path; nothing else moves.

Dependency category (per process/Deepening/deepening.md): in-process. No
port or adapter — every absorbed helper was pure computation over text.
The worker-pool is still an internal seam, used by the cleaner-async
path; not exposed.

Tests deleted: 0. Neither shallow module had unit tests of its own.
Filing-extractor-debug.ts was dead — no caller, no test, no .js shadow.
Filing-extractor-utils.ts was internal-only except for cleanHtmlContent;
that single export is now re-exported from the deep module.

ADRs respected:
- Mirrors ADR 0001 (inline filing-type registration): collapse parallel
  files defining one concept down to a single registry / module.
- Mirrors ADR 0002 (inline analysis-depth scoring): files extracted for
  imagined reuse but with one real consumer fold back into that consumer.
